app/utils/postgres.py
import os
from psycopg2.extras import RealDictCursor
import psycopg2
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

def get_db_config():
    """Get database configuration from environment variables"""
    DB_USER = os.getenv("POSTGRES_USER")
    DB_PASSWORD = os.getenv("POSTGRES_PASSWORD")
    DB_HOST = os.getenv("POSTGRES_HOST")
    DB_NAME = os.getenv("POSTGRES_DB")

    if not all([DB_USER, DB_PASSWORD, DB_HOST, DB_NAME]):
        raise ValueError("Missing required database environment variables")
    
    return {
        "dbname": DB_NAME,
        "user": DB_USER,
        "password": DB_PASSWORD,
        "host": DB_HOST
    }

def get_db_connection():
    """Create and return a database connection"""
    db_config = get_db_config()
    return psycopg2.connect(
        **db_config,
        cursor_factory=RealDictCursor
    ) 