pyproject.toml
[tool.poetry]
name = "home-api"
version = "0.1.0"
description = "Home API Service - A FastAPI service for home database"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
pydantic = "^2.5.1"
pydantic-settings = "^2.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
